Update pub_20240917_PHB.js
With the help of evanelric and TappyTap, I have shortened the class descriptions for the Monk and Paladin Classes and updated the Spell Creatures CompanionList condensing them into one listing instead of one for each spell. Updated Monk and Paladin class feature sheet integration, so they work more cohesively with the sheet's formatting.